+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Features
 
+* `[jest-validate]` Get rid of `jest-config` dependency
+  ([#6067](https://github.com/facebook/jest/pull/6067))
+* `[jest-validate]` Adds option to inject `deprecationEntries`
+  ([#6067](https://github.com/facebook/jest/pull/6067))
 * `[jest-snapshot]` [**BREAKING**] Concatenate name of test, optional snapshot
   name and count ([#6015](https://github.com/facebook/jest/pull/6015))
 * `[jest-runtime]` Allow for transform plugins to skip the definition process